Remove setuptools from install_require in setup.py
setuptools is considered unsafe to be set in install_require as all the
setup requirements for a project (python packaging pip related for
instance).

This PR remove it from the install_require. This also avoids some
conflicts or warning from other projects.

py.test change to have non return 0 if no tests are run, I removed the
assertion on the return code.